requirements.txt
flask==0.12.2
python-magic==0.4.15
